Refactor generic data generation to conftest
